api/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Telegram bot
python-telegram-bot==21.0

# HTTP client
httpx==0.27.0

# Data validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Environment
python-dotenv==1.0.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Markdown processing
markdown==3.5.1

# LLM Providers
anthropic>=0.18.0
openai>=1.10.0

# Vector Database
qdrant-client>=1.7.0

# Reranking
cohere>=4.40

# YAML frontmatter
pyyaml>=6.0

# Token counting for chunking
tiktoken>=0.5.0

# Testing (dev dependencies)
pytest==8.0.0
pytest-asyncio==0.23.3
pytest-cov==4.1.0
